requirements.txt
mysql-connector-python
pymysql
openai
streamlit
streamlit-chat
python-dotenv